+++ lua/issuehub/provider/redmine.lua
@@ -43,6 +43,7 @@ function Redmine:_ctx()
     name = self.name,
     base = self.base,
     http = self.http,
+    opts = self.opts,
     auth = function(token)
       return { headers = { ["X-Redmine-API-Key"] = token } }
     end,